Use f64 instead of float.
`float` has been removed from the language. SQLite documents their REAL type
as being an [f64](), so use that. Exposes a correctness bug where Rust
wouldn't pick f32 as the float type.

[f64] - http://sqlite.org/datatype3.html

+++ types.rs
@@ -104,7 +104,7 @@ impl to_str::ToStr for ResultCode {
 #[deriving(Eq)]
 pub enum BindArg {
   Text(~str),
-  Number(float),
+  Number(f64),
   Integer(int),
   Blob(~[u8]),
   Null,
